support includes in non-executable code blocks
