+++ src/core/handlers/base.ts
@@ -68,6 +68,8 @@ import { mappedStringFromFile } from "../mapped-text.ts";
 import { error } from "log/mod.ts";
 import { withCriClient } from "../cri/cri.ts";
 import { normalizePath } from "../path.ts";
+import { isBlockShortcode } from "../lib/parse-shortcode.ts";
+import { standaloneInclude } from "./include-standalone.ts";
 
 const handlers: Record<string, LanguageHandler> = {};
 
@@ -311,6 +313,34 @@ export function install(handler: LanguageHandler) {
   }
 }
 
+const processMarkdownIncludes = async (
+  newCells: MappedString[],
+  options: LanguageCellHandlerOptions,
+) => {
+  const includeHandler = makeHandlerContext({
+    ...options,
+  });
+  // search for include shortcodes in the cell content
+  for (let i = 0; i < newCells.length; ++i) {
+    const lines = mappedLines(newCells[i], true);
+    let foundShortcodes = false;
+    for (let j = 0; j < lines.length; ++j) {
+      const shortcode = isBlockShortcode(lines[j].value);
+      if (shortcode && shortcode.name === "include") {
+        foundShortcodes = true;
+        const param = shortcode.params[0];
+        if (!param) {
+          throw new Error("Include directive needs filename as a parameter");
+        }
+        lines[j] = await standaloneInclude(includeHandler.context, param);
+      }
+    }
+    if (foundShortcodes) {
+      newCells[i] = mappedConcat(lines);
+    }
+  }
+};
+
 export async function handleLanguageCells(
   options: LanguageCellHandlerOptions,
 ): Promise<{
@@ -438,6 +468,11 @@ export async function handleLanguageCells(
       }
     }
   }
+
+  // now handle the markdown content. This is necessary specifically for
+  // include shortcodes that can still be hiding inside of code blocks
+  await processMarkdownIncludes(newCells, options);
+
   return {
     markdown: mappedJoin(newCells, ""),
     results,
