+++ tests/docs/smoke-all/2023/12/01/file.r
@@ -1,0 +1,1 @@
+cat("Hello.")
